==> 01_analysis.R <==
library(tidyverse)
library(janitor)
library(fs)
library(lubridate)


#load precinct data from step 00
combined2018 <- readRDS("processed_data/combined2018_precincts.rds")

glimpse(combined2018)

combined2018$group_vote_ct_adj <- as.numeric(combined2018$group_vote_ct_adj)
combined2018$group_vote_ct_adj <- combined2018$group_vote_ct_adj - 1 #adjustment per documentation: see READ_ME file


#pull out only results for NC-09 congressional district
nc09_precincts18 <- combined2018 %>% 
  filter(contest_title == "US HOUSE OF REPRESENTATIVES DISTRICT 09")

#pull grand totals by candidate
nc09_precincts18 %>% 
  filter(!candidate_name %in% c("OVER VOTE", "UNDER VOTE")) %>% 
  group_by(candidate_name, candidate_party_lbl) %>% 
  summarise(cnt = n(), sum(group_vote_ct_adj))

#pull grand totals by candidate
nc09_grouped_precincts <- nc09_precincts18 %>% 
  filter(!candidate_name %in% c("OVER VOTE", "UNDER VOTE"),
         county_desc %in% c("UNION",
                              "MECKLENBURG",
                               "ANSON",
                               "RICHMOND",
                               "SCOTLAD",
                               "ROBESON",
                               "BLADEN",
                               "CUMBERLAND")
         ) %>% 
  group_by(county_desc, precinct_code, precinct_name, candidate_name, candidate_party_lbl) %>% 
  summarise(cnt = n(), num_votes = sum(group_vote_ct_adj))

nc09_grouped_precincts


#save results to files
write_csv(nc09_grouped_precincts, "processed_data/nc09_2018_grouped_precincts.csv")
saveRDS(nc09_grouped_precincts, "processed_data/nc09_2018_grouped_precincts.rds")




#### 2016 PRESIDENTIAL ####

#load saved version from step 00
nc2016_prez <- readRDS("processed_data/nc2016_prez.rds")

#filter to NC-09 counties only
nc09_prez16 <- nc2016_prez %>% 
  filter(county%in% c("UNION",
                                     "MECKLENBURG",
                                     "ANSON",
                                     "RICHMOND",
                                     "SCOTLAD",
                                     "ROBESON",
                                     "BLADEN",
                                     "CUMBERLAND"))

nc09_prez16
