==> 02_markdown_comparisons.Rmd <==
---
title: "NC-09 precinct-level comparisons"
author: "Aaron Kessler"
date: "9/12/2019"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(janitor)
library(lubridate)
library(reshape2)
library(writexl)
library(kableExtra)
# library(gt)
options(scipen = 999)


joined <- readRDS("processed_data/joined.rds")

robeson <- joined %>% 
  filter(county == "ROBESON")




```

### Let's start out with turnout: How many precincts saw more votes for each party in the special compared with the original midterms in November. How many saw less?

Democrats: Nearly every precinct saw votes *fall* from November
```{r echo=FALSE}

joined %>% 
  count(dem_updown_votes) %>% 
  arrange(desc(n)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```


The same was true for Republicans. 

```{r echo=FALSE}

joined %>% 
  count(gop_updown_votes) %>% 
  arrange(desc(n)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```

So we know this was a low-turnout election broadly and people who came out in their precincts in November didn't show up this time.  
  
  

### Now let's look at each party's **percent** of the vote in each precinct
#### Did Tuesday's margins go up or down since Nov?

Democratic
```{r echo=FALSE}

joined %>% 
  count(dem_updown) %>% 
  arrange(desc(n)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```


Republican
```{r echo=FALSE}

joined %>% 
  count(gop_updown) %>% 
  arrange(desc(n)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```

So we can see Republicans did significantly better. Even though turnout overall was lower, the Republican voters showed up more often at the precinct level, allowing the party to boost its margins this time.

#### Where did Democratic percent of the vote fall from Nov?

Let's see which counties had the largest number of precincts with lower Democratic margins this time compared with November's midterms
```{r echo=FALSE}

joined %>% 
  count(dem_updown, county) %>% 
  filter(dem_updown == "down") %>% 
  arrange(desc(n)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```

It's clear that Robeson County plays a key role in the Democratic struggles on Tues night. That much was known quite quickly, though let's dive in a little deeper now and see what the precincts tell us.

<!-- Where did Republican percent of the vote fall from Nov -->
<!-- ```{r echo=FALSE} -->

<!-- joined %>%  -->
<!--   count(gop_updown, county) %>%  -->
<!--   filter(gop_updown == "down") %>%  -->
<!--   arrange(desc(n)) %>%  -->
<!--   kable() %>%  -->
<!--   kable_styling(bootstrap_options = c("striped", "condensed"),  -->
<!--                 full_width = F, -->
<!--                 position = "left") -->

<!-- ``` -->

### Precincts that flipped parties on Tuesday

#### Let's see if Robeson features prominently or not among precincts that flipped from blue to red, or vice versa.

Total flipped precincts on Tues night show us there were 27 such precincts district-wide.
```{r echo=FALSE}

joined %>% 
  count(flip) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```


Looking by party, it's clear Republicans owned the night for flips 
```{r echo=FALSE}

joined %>% 
  filter(flip == "Y") %>% 
  count(flip, winner19) %>% 
  arrange(desc(n)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```

So now we also know that Republicans were much more successful on Tuesday night in flipping precincts.  
This matches up with our previous breakdowns above showing how Republicans gained greater shares of the vote in more precincts than Democrats.  
Now we know they were *also better a flipping them altogether* and winning precincts Democrats had carried in November.  

#### So which *counties* had most flips by party?
Republicans flipped a huge chunk (15!) in **Robeson County**
```{r echo=FALSE}

joined %>% 
  filter(flip == "Y") %>% 
  count(flip, winner19, county) %>% 
  arrange(desc(winner19), desc(n)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```


#### Diving deeper into Robeson County: How the Dems lost ground and why it was so devastating for Tuesday's results

Democrats lost vote share in nearly every Robeson precinct compared with 2018.
The table below shows the number of precincts the Dem share went up or down in *just Robeson*.
```{r echo=FALSE}

robeson %>% 
  count(dem_updown) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```


And in fact 33 is also the number of precincts Democrats won in 2018. This time? Not so much.

```{r echo=FALSE}

robeson %>% 
  count(winner18) %>% 
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"),
                full_width = F,
                position = "left")

robeson %>% 
  count(winner19) %>% 
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"),
                full_width = F,
                position = "left")



```

Here's what that looks like in terms of number of votes in those precincts. 2018 compared to 2019.
```{r echo=FALSE}

robeson %>% 
  group_by(winner18) %>% 
  summarise(n(), demvotes = sum(dem18), gopvotes = sum(gop18)) %>% 
  mutate(
    difference = demvotes - gopvotes
  ) %>% 
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"),
                full_width = F,
                position = "left")



robeson %>% 
  group_by(winner19) %>% 
  summarise(n(), demvotes = sum(dem19), gopvotes = sum(gop19)) %>% 
  mutate(
    difference = demvotes - gopvotes
  ) %>% 
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"),
                full_width = F,
                position = "left")


```


Democrats only had five precincts in Robeson with higher vote percentages than in November - and even those were low single digits.

By contrast, the losses were big and widespread. In 11 precincts, Democrats had double-digit losses in pct points compared with 2018 -- representing a loss of *more than 2,000 votes* since the midterms in just those precincts alone. (All but one of those precincts flipped from blue to red.)
```{r echo=FALSE}

robeson %>% 
  select(county, precinct, dem_change, dem_vote_change, flip) %>% 
  arrange(dem_change) %>% 
  head(11) %>% 
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"),
                full_width = F,
                position = "left") %>% 
  column_spec(3, bold = T)




```

Conversly, here are the biggest Republican wins in Robeson, and point growth from last time
```{r echo=FALSE}

robeson %>% 
  select(county, precinct, gop18pct, gop19pct) %>% 
  arrange(desc(gop19pct)) %>% 
  mutate(
    point_boost = gop19pct - gop18pct
  ) %>% 
  head(10) %>%
  kable() %>%
  kable_styling(bootstrap_options = c("striped", "condensed"),
                full_width = F,
                position = "left") %>%
  column_spec(4, bold = T)

```


There's even a more devastating way to look at Democratic performance in Robseon.  
That is to look at the net loss of votes for Democrats on Tues vs. November -- in other words, precinct by precinct, what was the vote spread between Dem and Rep votes back in November...and then what was the vote spread on Tuesday.  
  
When you look at the difference between those tallies, the net loss in Democratic advantage was staggering.  Democrats may have "won" Robeson County on Tues, but they lost the spread difference by more than 4,000 votes.
```{r echo=FALSE}

robeson %>% 
  summarise(sum(votespread_GAINdiff_dem)) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left")

```



### Vote-Spread Metrics


Where did Democrats *lose* the most in the vote number difference compared with Nov?
It's all about Robeson County.
```{r echo=FALSE}

joined %>% 
  select(county, precinct, votespread_18_dem, votespread_19_dem, votespread_GAINdiff_dem) %>% 
  arrange(votespread_GAINdiff_dem) %>% 
  filter(votespread_19_dem < 0) %>% 
  head(10) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left") %>% 
  column_spec(5, bold = T)

```



In areas where Democrats gained votes from Nov, where did they *gain* the most in vote number spread?
```{r echo=FALSE}

joined %>% 
  select(county, precinct, votespread_18_dem, votespread_19_dem, votespread_GAINdiff_dem) %>% 
  arrange(desc(votespread_GAINdiff_dem)) %>% 
  filter(votespread_19_dem > 0) %>% 
  head(10) %>% 
  kable() %>% 
  kable_styling(bootstrap_options = c("striped", "condensed"), 
                full_width = F,
                position = "left") %>% 
  column_spec(5, bold = T)

```
